Move MIME type checking to the middleware
